pyproject.toml
[project]
name = "fastapi-backend-template"
version = "0.1.0"
description = "Reusable FastAPI backend template with auth, RBAC, Celery, Redis, and object storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=21.2.0",
    "python-multipart>=0.0.20",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.16.0",

    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Authentication & security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<5.0.0",
    "email-validator>=2.0.0",

    # Background tasks
    "celery>=5.4.0",
    "redis>=5.0.1",
    "flower>=2.0.1",

    # CLI utilities
    "typer>=0.12.0",
    "rich>=13.0.0",

    # File handling / object storage
    "boto3>=1.42.49",
    "Pillow>=11.0.0",
    "python-magic>=0.4.27",

    # Data processing (optional — remove if not needed)
    "pandas>=3.0.0",
    "openpyxl>=3.1.5",

    # Monitoring
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]
